feat: 코드 분리 및 요구사항과 구현 순서 기록
