+++ README.md
@@ -6,13 +6,11 @@
 
 - [x] 자동차에 이름을 부여할 수 있다. 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
 - [x] 자동차 이름은 쉼표(,)를 기준으로 구분하며 이름은 5자 이하만 가능하다.
-  - [x] 자동차는 최소 2개를 입력 받는다.
 - [x] 자동차 경주는 5회로 고정하여 진행한다.
 - [x] 전진하는 조건은 0에서 9 사이에서 무작위 값을 구한 후 무작위 값이 4 이상일 경우이다.
 - [x] 자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한 명 이상일 수 있다.
 - [x] 우승자가 여러 명일 경우 쉼표(,)를 이용하여 구분한다.
 - [x] 사용자가 잘못된 입력 값을 작성한 경우 프로그램을 종료한다.
-  - [x] 에러 메세지 없이 바로 return 한다?
 
 ```
 경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).
