+++ src/domain/isNameValidation.js
@@ -1,0 +1,5 @@
+export const isNameValidation = (name) => {
+  if (name.split(",").every((item) => item.length > 6)) return false;
+  if (name.split(",").length === 1) return false;
+  return true;
+};
