+++ src/utils/getRandomNumber.js
@@ -1,0 +1,3 @@
+export const getRandomNumber = () => {
+  return Math.floor(Math.random() * 10);
+};
